Reuse `groupBy` in validation rules
Follow up to #3208
