+++ src/validation/rules/UniqueArgumentNamesRule.ts
@@ -1,4 +1,8 @@
+import { groupBy } from '../../jsutils/groupBy';
+
 import { GraphQLError } from '../../error/GraphQLError';
+
+import type { ArgumentNode } from '../../language/ast';
 import type { ASTVisitor } from '../../language/visitor';
 
 import type { ASTValidationContext } from '../ValidationContext';
@@ -14,27 +18,28 @@ import type { ASTValidationContext } from '../ValidationContext';
 export function UniqueArgumentNamesRule(
   context: ASTValidationContext,
 ): ASTVisitor {
-  let knownArgNames = Object.create(null);
   return {
-    Field() {
-      knownArgNames = Object.create(null);
-    },
-    Directive() {
-      knownArgNames = Object.create(null);
-    },
-    Argument(node) {
-      const argName = node.name.value;
-      if (knownArgNames[argName]) {
+    Field: checkArgUniqueness,
+    Directive: checkArgUniqueness,
+  };
+
+  function checkArgUniqueness(parentNode: {
+    arguments?: ReadonlyArray<ArgumentNode>;
+  }) {
+    // istanbul ignore next (See: 'https://github.com/graphql/graphql-js/issues/2203')
+    const argumentNodes = parentNode.arguments ?? [];
+
+    const seenArgs = groupBy(argumentNodes, (arg) => arg.name.value);
+
+    for (const [argName, argNodes] of seenArgs) {
+      if (argNodes.length > 1) {
         context.reportError(
           new GraphQLError(
             `There can be only one argument named "${argName}".`,
-            [knownArgNames[argName], node.name],
+            argNodes.map((node) => node.name),
           ),
         );
-      } else {
-        knownArgNames[argName] = node.name;
       }
-      return false;
-    },
-  };
+    }
+  }
 }
